change function selectors
Some services that inspect bytecode to determine the token type have falsely identified the base contract as an ERC721 due to the existence of certain function selectors in the fallback function. This change updates the function selectors for communication between the mirror contract and the base contract to avoid any overlap with ERC721 function selectors. It has been tested in a production environment and the services that were previously picking up the base ERC20 tokens as ERC721 no longer identify them as such.